[WIP] Scratch of the function to update velocity model #1
Use flat layers model and scratch the function to update
this model. It is an easy way to start because it does not
need to calculate z coordinates for all interface, you just
use the first point as the interface z coordinate for every x.

With this function, fill in the vel matrix by column