Dynamic ray tracing for the first layer #4
The program calculation of RNIP is working for the first layer. Although
it is not working for the second one, because it needs Snell's law to
be used to RNIP trasmission through the interface.

The propagation law is used to check if the RNIP value is correct,
it is just vi*dt, the length of the normal ray throughout the layer