Velocity inversion working #4
Code refactoring of the experiment and parameters update. Now the
inversion seems to work properly. Unfortunately the interface shape
is a function of the number of points considered in 'pick.txt'. For
13 points for each interface the algorithm can get interface shape, it
is not the same if we use 22 points or more.

The velocity is well inverted with a small error.